+++ rust/crd/src/affinity.rs
@@ -20,6 +20,9 @@ pub fn get_affinity(
             .filter_map(|catalog| match &catalog.spec.connector {
                 TrinoCatalogConnector::Hive(hive) => Some(&hive.metastore.config_map),
                 TrinoCatalogConnector::Iceberg(iceberg) => Some(&iceberg.metastore.config_map),
+                TrinoCatalogConnector::DeltaLake(delta_lake) => {
+                    Some(&delta_lake.metastore.config_map)
+                }
                 TrinoCatalogConnector::BlackHole(_)
                 | TrinoCatalogConnector::Generic(_)
                 | TrinoCatalogConnector::GoogleSheet(_)
@@ -43,6 +46,9 @@ pub fn get_affinity(
                 }
                 TrinoCatalogConnector::Iceberg(iceberg) => {
                     iceberg.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
+                }
+                TrinoCatalogConnector::DeltaLake(delta_lake) => {
+                    delta_lake.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
                 }
                 TrinoCatalogConnector::BlackHole(_)
                 | TrinoCatalogConnector::Generic(_)
@@ -86,11 +92,10 @@ mod tests {
 
     use crate::TrinoCluster;
     use stackable_operator::{
-        commons::affinity::{StackableAffinity, StackableNodeSelector},
+        commons::affinity::StackableAffinity,
         k8s_openapi::{
             api::core::v1::{
-                NodeAffinity, NodeSelector, NodeSelectorRequirement, NodeSelectorTerm, PodAffinity,
-                PodAffinityTerm, PodAntiAffinity, WeightedPodAffinityTerm,
+                PodAffinity, PodAffinityTerm, PodAntiAffinity, WeightedPodAffinityTerm,
             },
             apimachinery::pkg::apis::meta::v1::LabelSelector,
         },
@@ -401,123 +406,6 @@ mod tests {
                 }),
                 node_affinity: None,
                 node_selector: None,
-            }
-        );
-    }
-
-    #[test]
-    fn test_affinity_legacy_node_selector() {
-        let input = r#"
-        apiVersion: trino.stackable.tech/v1alpha1
-        kind: TrinoCluster
-        metadata:
-          name: simple-trino
-        spec:
-          image:
-            productVersion: "428"
-          clusterConfig:
-            catalogLabelSelector:
-              matchLabels:
-                trino: simple-trino
-          coordinators:
-            roleGroups:
-              default:
-                replicas: 1
-                selector:
-                  matchLabels:
-                    disktype: ssd
-                  matchExpressions:
-                    - key: topology.kubernetes.io/zone
-                      operator: In
-                      values:
-                        - antarctica-east1
-                        - antarctica-west1
-          workers:
-            roleGroups:
-              default:
-                replicas: 1
-        "#;
-        let trino: TrinoCluster = serde_yaml::from_str(input).expect("illegal test input");
-        let merged_config = trino
-            .merged_config(
-                &TrinoRole::Coordinator,
-                &TrinoRole::Coordinator.rolegroup_ref(&trino, "default"),
-                &[],
-            )
-            .unwrap();
-
-        assert_eq!(
-            merged_config.affinity,
-            StackableAffinity {
-                pod_affinity: Some(PodAffinity {
-                    preferred_during_scheduling_ignored_during_execution: Some(vec![
-                        WeightedPodAffinityTerm {
-                            pod_affinity_term: PodAffinityTerm {
-                                label_selector: Some(LabelSelector {
-                                    match_expressions: None,
-                                    match_labels: Some(BTreeMap::from([
-                                        ("app.kubernetes.io/name".to_string(), "trino".to_string(),),
-                                        (
-                                            "app.kubernetes.io/instance".to_string(),
-                                            "simple-trino".to_string(),
-                                        ),
-                                    ]))
-                                }),
-                                namespace_selector: None,
-                                namespaces: None,
-                                topology_key: "kubernetes.io/hostname".to_string(),
-                            },
-                            weight: 20
-                        }
-                    ]),
-                    required_during_scheduling_ignored_during_execution: None,
-                }),
-                pod_anti_affinity: Some(PodAntiAffinity {
-                    preferred_during_scheduling_ignored_during_execution: Some(vec![
-                        WeightedPodAffinityTerm {
-                            pod_affinity_term: PodAffinityTerm {
-                                label_selector: Some(LabelSelector {
-                                    match_expressions: None,
-                                    match_labels: Some(BTreeMap::from([
-                                        ("app.kubernetes.io/name".to_string(), "trino".to_string(),),
-                                        (
-                                            "app.kubernetes.io/instance".to_string(),
-                                            "simple-trino".to_string(),
-                                        ),
-                                        (
-                                            "app.kubernetes.io/component".to_string(),
-                                            "coordinator".to_string(),
-                                        )
-                                    ]))
-                                }),
-                                namespace_selector: None,
-                                namespaces: None,
-                                topology_key: "kubernetes.io/hostname".to_string(),
-                            },
-                            weight: 70
-                        }
-                    ]),
-                    required_during_scheduling_ignored_during_execution: None,
-                }),
-                node_affinity: Some(NodeAffinity {
-                    preferred_during_scheduling_ignored_during_execution: None,
-                    required_during_scheduling_ignored_during_execution: Some(NodeSelector {
-                        node_selector_terms: vec![NodeSelectorTerm {
-                            match_expressions: Some(vec![NodeSelectorRequirement {
-                                key: "topology.kubernetes.io/zone".to_string(),
-                                operator: "In".to_string(),
-                                values: Some(vec![
-                                    "antarctica-east1".to_string(),
-                                    "antarctica-west1".to_string()
-                                ]),
-                            }]),
-                            match_fields: None,
-                        }]
-                    }),
-                }),
-                node_selector: Some(StackableNodeSelector {
-                    node_selector: BTreeMap::from([("disktype".to_string(), "ssd".to_string())])
-                }),
             }
         );
     }