+++ rust/operator-binary/src/catalog/commons.rs
@@ -19,6 +19,7 @@ use stackable_trino_crd::{
 use super::{
     config::CatalogConfig,
     from_trino_catalog_error::{
+        CreateS3CredentialsSecretOperatorVolumeSnafu, CreateS3TLSSecretOperatorVolumeSnafu,
         FailedToGetDiscoveryConfigMapDataKeySnafu, FailedToGetDiscoveryConfigMapDataSnafu,
         FailedToGetDiscoveryConfigMapSnafu, ObjectHasNoNamespaceSnafu, ResolveS3ConnectionDefSnafu,
         S3TlsNoVerificationNotSupportedSnafu,
@@ -104,7 +105,11 @@ impl ExtendCatalogConfig for S3ConnectionDef {
             let volume_mount_path = format!("{S3_SECRET_DIR_NAME}/{catalog_name}/{secret_class}");
             catalog_config.volumes.push(
                 VolumeBuilder::new(&volume_name)
-                    .ephemeral(SecretOperatorVolumeSourceBuilder::new(&secret_class).build())
+                    .ephemeral(
+                        SecretOperatorVolumeSourceBuilder::new(&secret_class)
+                            .build()
+                            .context(CreateS3CredentialsSecretOperatorVolumeSnafu)?,
+                    )
                     .build(),
             );
             catalog_config
@@ -137,7 +142,11 @@ impl ExtendCatalogConfig for S3ConnectionDef {
                         format!("{STACKABLE_MOUNT_CLIENT_TLS_DIR}/{catalog_name}/{secret_class}");
                     catalog_config.volumes.push(
                         VolumeBuilder::new(&volume_name)
-                            .ephemeral(SecretOperatorVolumeSourceBuilder::new(secret_class).build())
+                            .ephemeral(
+                                SecretOperatorVolumeSourceBuilder::new(secret_class)
+                                    .build()
+                                    .context(CreateS3TLSSecretOperatorVolumeSnafu)?,
+                            )
                             .build(),
                     );
                     catalog_config
@@ -172,7 +181,7 @@ impl ExtendCatalogConfig for HdfsConnection {
         let hdfs_site_dir = format!("{CONFIG_DIR_NAME}/catalog/{catalog_name}/hdfs-config");
         catalog_config.add_property(
             "hive.config.resources",
-            format!("{hdfs_site_dir}/hdfs-site.xml"),
+            format!("{hdfs_site_dir}/core-site.xml,{hdfs_site_dir}/hdfs-site.xml"),
         );
 
         let volume_name = format!("{catalog_name}-hdfs");