+++ CHANGELOG.md
@@ -7,8 +7,17 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Various documentation of the CRD ([#510]).
+- Helm: support labels in values.yaml ([#528]).
+- Delta Lake connector ([#531]).
+
+### Fixed
+
+- Add `core-site.xml` when configuring HDFS connection ([#526]).
 
 [#510]: https://github.com/stackabletech/trino-operator/pull/510
+[#526]: https://github.com/stackabletech/trino-operator/pull/526
+[#528]: https://github.com/stackabletech/trino-operator/pull/528
+[#531]: https://github.com/stackabletech/trino-operator/pull/531
 
 ## [23.11.0] - 2023-11-24